vt: support 24-bit color
SGR 38/48 only had the ";5;N" branch. On "38;2;R;G;B" the condition failed
without advancing the index, so the remaining parameters were read as
standalone SGR codes: the 2 became "dim" and the RGB components landed in
the 30-37, 40-47 and 90-97 ranges, i.e. they were rendered as ANSI colors.
Both separators are accepted now -- ECMA-48 splits subparameters with ':',
and the two layouts put the RGB at different offsets, since the colon form
carries a color space in between.

Cells store colors as int32_t, and the 256-color palette is folded into RGB
so that one representation reaches the pair allocator.

color2palette was a dense (COLORS+2)^2 matrix of shorts. With direct color
that allocation asks for 563 TB, so dvtm could not even start under a
direct-color terminfo -- alloc_pair() keeps that table instead, and the
round-robin recycling is gone with it. Pairs above 255 do not fit the 8 bits
of COLOR_PAIR(), so the pair now travels through the extended argument of
wcolor_set().

The terminfo advertises Tc/setrgbf rather than colors#0x1000000: declaring
16.7M colors to children breaks zsh's color arithmetic -- it emits a bare
"ESC[38m" and drops the backspaces it uses to redraw the prompt, which
duplicates every character typed.

+++ dvtm.c
@@ -91,7 +91,7 @@ typedef struct {
 	short bg;
 	short fg256;
 	short bg256;
-	short pair;
+	int pair;
 } Color;
 
 typedef struct {
